requirements_4_image.txt
apache-airflow-providers-postgres==5.6.0